refactor(codegen): extract std::io::sockopt::* into stdlib/sockopt.rs (Round 1.9)
One generic getter lifted into SockoptStdlib<'ctx>; the ~30 named
constants (IPPROTO_*, IP_*, SO_*, SOL_SOCKET) dispatch to it from
codegen.rs. SOCKOPT_NAMES const stays in codegen.rs (also consulted
at declare-builtins time).

codegen.rs: 43544 -> 43508 (-36 lines).
