+++ crates/hale-codegen/src/stdlib/mod.rs
@@ -15,5 +15,6 @@ pub(crate) mod env;
 pub(crate) mod math;
 pub(crate) mod process;
 pub(crate) mod rand;
+pub(crate) mod sockopt;
 pub(crate) mod text;
 pub(crate) mod time;
